fix implicit bool conversions

+++ power_grid_model_c/power_grid_model_c/src/meta_data.cpp
@@ -85,4 +85,4 @@ PGM_Idx PGM_meta_attribute_ctype(PGM_Handle* /*unused*/, PGM_MetaAttribute const
 size_t PGM_meta_attribute_offset(PGM_Handle* /*unused*/, PGM_MetaAttribute const* attribute) {
     return attribute->offset;
 }
-int PGM_is_little_endian(PGM_Handle* /*unused*/) { return meta_data::is_little_endian(); }
+int PGM_is_little_endian(PGM_Handle* /*unused*/) { return static_cast<int>(meta_data::is_little_endian()); }
